investment-bot/requirements.txt
`requirements.txt`:

discord.py
python-dotenv
-U discord.py


Installera med:
```bash
pip install -r requirements.txt
```